connections-service/requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
httpx>=0.27.0
python-telegram-bot>=21.0
discord.py>=2.3.0
